requirements.txt
fastapi  
uvicorn[standard]  
sqlalchemy  
psycopg2-binary  
python-dotenv  
pydantic-settings  
passlib[bcrypt]  
python-multipart
pytest
httpx
scipy
numpy
apscheduler
aiohttp
fastapi-mail
tiktoken
openai
alembic
redis
hiredis
slowapi
